refactor: Extracted function (createSignature)
