+++ src/steps/create-signatures.ts
@@ -4,45 +4,8 @@ import { join } from 'node:path';
 import { createFiles } from '@codemod-utils/files';
 
 import type { Context, Options } from '../types/index.js';
-import {
-  getComponentFilePath,
-  type TransformedEntityName,
-  transformEntityName,
-} from '../utils/files.js';
-import {
-  getBaseComponentName,
-  passSignatureToBaseComponent,
-  updateReferences,
-} from './create-signatures/index.js';
-
-type Data = {
-  entity: TransformedEntityName;
-};
-
-function createSignature(file: string, data: Data): string {
-  const baseComponentName = getBaseComponentName(file);
-
-  if (baseComponentName === undefined) {
-    return file;
-  }
-
-  // eslint-disable-next-line prefer-const
-  let { interfaceName, newFile } = passSignatureToBaseComponent(file, {
-    baseComponentName,
-    data,
-  });
-
-  if (interfaceName === undefined) {
-    return newFile;
-  }
-
-  ({ newFile } = updateReferences(newFile, {
-    data,
-    interfaceName,
-  }));
-
-  return newFile;
-}
+import { getComponentFilePath, transformEntityName } from '../utils/files.js';
+import { createSignature } from './create-signatures/index.js';
 
 export function createSignatures(context: Context, options: Options): void {
   const { projectRoot } = options;
