+++ src/steps/create-signatures/index.ts
@@ -1,3 +1,1 @@
-export * from './get-base-component-name.js';
-export * from './pass-signature-to-base-component.js';
-export * from './update-references.js';
+export * from './create-signature.js';
